% Auto-generated: do not edit by hand
\name{''DashDraggableGrid}

\alias{''DashDraggableGrid}

\title{DashDraggableGrid component}

\description{
GridComponent is an dash component. It takes a property, `id`, and children to displays it.
}

\usage{
''DashDraggableGrid(children=NULL, id=NULL, label=NULL, editable=NULL,
position=NULL, value=NULL)
}

\arguments{
\item{children}{List of a list of or a singular dash component, string or numbers | a list of or a singular dash component, string or number. The children components displayed inside the grid.}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{label}{Character. A label that will be printed when this component is rendered.}

\item{editable}{Logical. The editing property displayed inside the grid.}

\item{position}{Unnamed list. The layout of the  components displayed inside the grid.}

\item{value}{Character. The value displayed in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

